+++ CHANGELOG.md
@@ -20,6 +20,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Initialize map with `len(keys)` in `NewAllowKeysFilter` and `NewDenyKeysFilter` to avoid unnecessary allocations in `go.opentelemetry.io/otel/attribute`. (#6455)
 - `go.opentelemetry.io/otel/log/logtest` is now a separate Go module. (#6465)
 - `go.opentelemetry.io/otel/sdk/log/logtest` is now a separate Go module. (#6466)
+- Refactor `go.opentelemetry.io/otel/sdk/log` to use `Record` pointers.
+  Changes `Exporter` interface to accept `[]*Record` instead of `[]Record`. (#6466)
+- Refactor `go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploggrpc` to use `Record` pointers. (#6466)
+- Refactor `go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp` to use `Record` pointers. (#6466)
+- Refactor `go.opentelemetry.io/otel/exporters/otlp/stdout/stdoutlog` to use `Record` pointers. (#6466)
 
 ### Fixes